Change edit account from dialog to activity
