+++ example/src/main/res/layout/activity_account_edit.xml
@@ -5,7 +5,6 @@
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:orientation="vertical"
         android:padding="@dimen/element_spacing_large">
 
         <TextView
